+++ src/Sniffs/Naming/InterfaceNameSniff.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 /*
  * Originally part of https://github.com/symplify/symplify