+++ tests/Fixer/SpecifyArgSeparatorFixerTest.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace Lmc\CodingStandard\Fixer;
 
@@ -8,14 +10,17 @@ use PHPUnit\Framework\TestCase;
 class SpecifyArgSeparatorFixerTest extends TestCase
 {
     /**
-     * @test
      * @dataProvider provideFiles
      */
-    public function shouldFixCode(string $inputFile, string $expectedOutputFile): void
+    public function testShouldFixCode(string $inputFile, string $expectedOutputFile): void
     {
         $fixer = new SpecifyArgSeparatorFixer();
         $fileInfo = new \SplFileInfo(__DIR__ . '/Fixtures/' . $inputFile);
-        $tokens = Tokens::fromCode(file_get_contents($fileInfo->getRealPath()));
+        $fileContents = file_get_contents($fileInfo->getRealPath());
+        if ($fileContents === false) {
+            $this->fail('Could not read file ' . $fileInfo->getRealPath());
+        }
+        $tokens = Tokens::fromCode($fileContents);
 
         $fixer->fix($fileInfo, $tokens);
 
@@ -26,9 +31,9 @@ class SpecifyArgSeparatorFixerTest extends TestCase
     }
 
     /**
-     * @return array[]
+     * @return array<string, array{string, string}>
      */
-    public function provideFiles(): array
+    public static function provideFiles(): array
     {
         return [
             'Correct file should not be changed' => ['Correct.php.inc', 'Correct.php.inc'],